Move migration compatibility behavior to per-adapter modules
Adapter-specific migration compatibility logic was scattered across
a single monolithic Compatibility class with adapter_name string
comparisons. This made it hard to maintain and impossible for
third-party adapters to extend without monkey-patching.

Each built-in adapter now defines its own MigrationCompatibility
module and returns it from the new
`migration_compatibility_module_for` hook. Third-party adapters
(e.g. cockroachdb) can override this hook to ship per-version
legacy behavior without patching Rails internals.

The hook is registered on Mysql2Adapter and TrilogyAdapter (not
AbstractMysqlAdapter) so third-party adapters subclassing the
abstract class are not implicitly opted in.

The assembled module is included on the topmost user-defined class
in the migration ancestry, preserving the pre-refactor lookup order
where user overrides run ahead of adapter-specific code.
Migrator#apply_adapter_compatibility_module covers migrations that
override #migrate directly. Module caching uses
ObjectSpace::WeakMap to avoid leaking anonymous/reloaded classes.
